// src/types.ts
export type Pet = {
  id: string;
  markno: string;
  name: string;
  name2: string;
  /**
   * 精力
   */
  jingli: number;
  /**
   * 攻击
   */
  gongji: number;
  /**
   * 防御
   */
  fangyu: number;
  /**
   * 魔攻
   */
  mougong: number;
  /**
   * 魔防
   */
  moufang: number;
  /**
   * 速度
   */
  sudu: number;
  total: number;
  img: string;
  href: string;
  href2: string;
  href3: string;
  href4: string;
  arcatt: number;
  typeid: number;
  title: string;
  title2: string | null;
  series: string;
};